demo/requirements.txt
asgiref==3.5.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.0.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
mssql-django==1.1.3
packaging==21.3
pymongo==4.1.1
pyodbc==4.0.32
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
sqlparse==0.4.2
wrapt==1.14.1
